lint: keep main() inside the cpplint size limit and spell gcount() as int64_t (#143)
The option-validation added for --count-resident-tiles pushed main() past the
500 non-comment-line limit and cpplint rejects the C type long, so the
non-negative check moves to a free function beside the other option helpers and
the partial-record message casts to int64_t. Also re-indents the replay lambda
to uncrustify's reading of its new first line. No behaviour change.

+++ cube_bathymetry/src/recon.cpp
@@ -23,6 +23,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstdint>
 #include <filesystem>
 #include <iostream>
 #include <limits>
@@ -230,7 +231,7 @@ uint64_t ReconCollector::forEachSpilled(const std::function<void(const GeoSoundi
   if (in.gcount() != 0) {
     throw std::runtime_error(
             "recon: spill file " + spillPath() + " ends in a partial record (" +
-            std::to_string(static_cast<long long>(in.gcount())) + " of " +
+            std::to_string(static_cast<int64_t>(in.gcount())) + " of " +
             std::to_string(sizeof(record)) + " bytes) after " +
             std::to_string(replayed) + " soundings; the spill is truncated");
   }
